+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Minor
 
+- IconButton: Add new `disabled` prop and stylings to `IconButton` component (#517)
+
 ### Patch
 
 </details>